+++ README.md
@@ -1,1 +1,4 @@
-# java-fx-mvc-examples
+# MVC calculator example
+
+### Illustrates how a project can both implement a Command-Line-Interface and Graphical-User-Interface using the MVC pattern
+
