+++ src/main/java/com/example/MVCCalculator/view/CalculatorView.java
@@ -1,4 +1,4 @@
-package MVCCalculator.view;
+package com.example.MVCCalculator.view;
 
 import javafx.event.EventHandler;
